move control policies
